requirements.txt
cycler==0.10.0
django-developer-panel==0.1.2
kiwisolver==1.1.0
matplotlib==3.1.2
mysql-connector-python==8.0.19
numpy==1.18.1
pandas==1.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
virtualenv==16.7.9
